Change current value type to QString
I decided that using a QString to store the current value makes more
sense than using a double. Implementing the calculator this way causes
us to evaluate the expression at the very end (when I implement that)
rather than as we go, but I think it will be easier to work with,
especially when we get to dealing with order of operations and
possibly the addition of parentheses in the future.