Add connections to operator buttons, equals button
I added connections to the operator buttons that causes the symbol
to appear in the displayed value when any operator buttons are
pressed. I also added an equals button, but did not incorporate any
of the logic to actually evaluate the current expression. Pressing
the equal button now will simply set the current value to 0.