Implement evaluation function
This commit includes several changes. Most importantly, a basic,
briefly tested version of the evaluation function has been
implemented. I removed the capability of our calculator to
support expressions like 2(1 + 3), where no operator sits between
the number and the parentheses. I also removed the tests that
check for this. The evaluation function passes all the current
tests, but I will be adding more to confirm that it is correct.